requirements.txt
flask
langdetect
openai
faiss-cpu
scikit-learn
numpy
